refactor(app): migrate MoviesSearchScreenTest
