--- shared-ui/src/androidMain/kotlin/com/vlohachov/shared/ui/screen/search/MoviesSearchScreen.android.kt
@@ -1,231 +1,0 @@
-package com.vlohachov.shared.ui.screen.search
-
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.animateColorAsState
-import androidx.compose.animation.core.FastOutLinearInEasing
-import androidx.compose.animation.core.Spring
-import androidx.compose.animation.core.spring
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.animation.scaleIn
-import androidx.compose.animation.scaleOut
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.ime
-import androidx.compose.foundation.layout.navigationBarsPadding
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.GridItemSpan
-import androidx.compose.foundation.lazy.grid.LazyGridState
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.rounded.ArrowBack
-import androidx.compose.material.icons.rounded.Clear
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SearchBar
-import androidx.compose.material3.SearchBarDefaults
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBarColors
-import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.material3.TopAppBarScrollBehavior
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.SideEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.derivedStateOf
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.lerp
-import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.platform.SoftwareKeyboardController
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.unit.dp
-import com.vlohachov.shared.core.collectAsLazyPagingItems
-import com.vlohachov.shared.domain.model.movie.Movie
-import com.vlohachov.shared.ui.component.bar.ErrorBar
-import com.vlohachov.shared.ui.component.button.ScrollToTop
-import com.vlohachov.shared.ui.component.movie.MoviesPaginatedGrid
-import moviespot.shared_ui.generated.resources.Res
-import moviespot.shared_ui.generated.resources.clear
-import moviespot.shared_ui.generated.resources.navigate_back
-import moviespot.shared_ui.generated.resources.search
-import org.jetbrains.compose.resources.stringResource
-
-private const val VISIBLE_ITEMS_THRESHOLD = 3
-private const val COLOR_TRANSITION_FRACTION_THRESHOLD = 0.01f
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-internal actual fun MoviesSearch(
-    onBack: () -> Unit,
-    onMovieDetails: (Movie) -> Unit,
-    viewModel: MoviesSearchViewModel,
-    gridState: LazyGridState,
-    snackbarHostState: SnackbarHostState,
-    scrollBehavior: TopAppBarScrollBehavior,
-    keyboardController: SoftwareKeyboardController?,
-) {
-    val showScrollToTop by remember { derivedStateOf { gridState.firstVisibleItemIndex > VISIBLE_ITEMS_THRESHOLD } }
-    ErrorBar(
-        error = viewModel.error.collectAsState().value,
-        snackbarHostState = snackbarHostState,
-        onDismissed = viewModel::onErrorConsumed,
-    )
-    Scaffold(
-        modifier = Modifier
-            .fillMaxSize()
-            .nestedScroll(connection = scrollBehavior.nestedScrollConnection),
-        topBar = {
-            SearchBar(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(vertical = 8.dp, horizontal = 16.dp),
-                value = viewModel.search.collectAsState(initial = "").value,
-                onValueChange = viewModel::onSearch,
-                onBack = onBack,
-                onClear = viewModel::onClear,
-                scrollBehavior = scrollBehavior,
-            )
-        },
-        floatingActionButton = {
-            ScrollToTop(
-                modifier = Modifier.navigationBarsPadding(),
-                visible = showScrollToTop,
-                gridState = gridState
-            )
-        },
-        snackbarHost = {
-            SnackbarHost(
-                modifier = Modifier
-                    .testTag(tag = SearchMoviesDefaults.ContentErrorTestTag)
-                    .navigationBarsPadding(),
-                hostState = snackbarHostState,
-            )
-        },
-        contentWindowInsets = WindowInsets.ime
-    ) { paddingValues ->
-        Content(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(paddingValues = paddingValues),
-            viewModel = viewModel,
-            gridState = gridState,
-            onMovieClick = { movie ->
-                keyboardController?.hide()
-                onMovieDetails(movie)
-            },
-        )
-    }
-}
-
-@Composable
-private fun Content(
-    modifier: Modifier,
-    viewModel: MoviesSearchViewModel,
-    gridState: LazyGridState,
-    onMovieClick: (Movie) -> Unit,
-) {
-    Box(modifier = modifier) {
-        MoviesPaginatedGrid(
-            modifier = Modifier.fillMaxSize(),
-            state = gridState,
-            columns = GridCells.Fixed(count = 3),
-            movies = viewModel.movies.collectAsLazyPagingItems(),
-            onClick = onMovieClick,
-            onError = viewModel::onError,
-            progress = {
-                item(span = { GridItemSpan(currentLineSpan = 3) }) {
-                    Box(
-                        modifier = Modifier
-                            .testTag(tag = SearchMoviesDefaults.ContentLoadingTestTag)
-                            .fillMaxWidth(),
-                        contentAlignment = Alignment.Center,
-                    ) {
-                        CircularProgressIndicator()
-                    }
-                }
-            }
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun SearchBar(
-    value: String,
-    onValueChange: (value: String) -> Unit,
-    onBack: () -> Unit,
-    onClear: () -> Unit,
-    modifier: Modifier = Modifier,
-    scrollBehavior: TopAppBarScrollBehavior,
-    appBarColors: TopAppBarColors = TopAppBarDefaults.centerAlignedTopAppBarColors(),
-) {
-    val heightOffsetLimit = with(LocalDensity.current) { -64.dp.toPx() }
-
-    SideEffect {
-        if (scrollBehavior.state.heightOffsetLimit != heightOffsetLimit) {
-            scrollBehavior.state.heightOffsetLimit = heightOffsetLimit
-        }
-    }
-
-    val colorTransitionFraction = scrollBehavior.state.overlappedFraction
-    val fraction = if (colorTransitionFraction > COLOR_TRANSITION_FRACTION_THRESHOLD) 1f else 0f
-    val appBarContainerColor by animateColorAsState(
-        targetValue = lerp(
-            start = appBarColors.containerColor,
-            stop = appBarColors.scrolledContainerColor,
-            fraction = FastOutLinearInEasing.transform(fraction)
-        ),
-        animationSpec = spring(stiffness = Spring.StiffnessMediumLow),
-        label = "anim_app_bar_container_color"
-    )
-    Surface(color = appBarContainerColor) {
-        SearchBar(
-            modifier = modifier.testTag(tag = SearchMoviesDefaults.SearchFieldTestTag),
-            query = value,
-            onQueryChange = onValueChange,
-            active = false,
-            onActiveChange = { },
-            onSearch = { },
-            placeholder = { Text(text = stringResource(resource = Res.string.search)) },
-            leadingIcon = {
-                IconButton(onClick = onBack) {
-                    Icon(
-                        imageVector = Icons.AutoMirrored.Rounded.ArrowBack,
-                        contentDescription = stringResource(resource = Res.string.navigate_back),
-                    )
-                }
-            },
-            trailingIcon = {
-                AnimatedVisibility(
-                    visible = value.isNotBlank(),
-                    enter = fadeIn() + scaleIn(),
-                    exit = fadeOut() + scaleOut(),
-                ) {
-                    IconButton(
-                        modifier = Modifier.testTag(tag = SearchMoviesDefaults.ClearSearchFieldTestTag),
-                        onClick = onClear,
-                    ) {
-                        Icon(
-                            imageVector = Icons.Rounded.Clear,
-                            contentDescription = stringResource(resource = Res.string.clear),
-                        )
-                    }
-                }
-            },
-            content = {},
-            colors = SearchBarDefaults.colors(containerColor = MaterialTheme.colorScheme.surfaceVariant)
-        )
-    }
-}
