+++ shared-ui/src/commonMain/kotlin/com/vlohachov/shared/core/LazyPagingItems.kt
@@ -54,8 +54,11 @@ internal class LazyPagingItems<T : Any> internal constructor(
      */
     private val pagingDataPresenter = object : PagingDataPresenter<T>(
         mainContext = mainDispatcher,
-        cachedPagingData =
-        if (flow is SharedFlow<PagingData<T>>) flow.replayCache.firstOrNull() else null
+        cachedPagingData = if (flow is SharedFlow<PagingData<T>>) {
+            flow.replayCache.firstOrNull()
+        } else {
+            null
+        }
     ) {
         override suspend fun presentPagingDataEvent(
             event: PagingDataEvent<T>,
